requirements.txt
python-dotenv==0.21.0
gspread==5.6.2
requests==2.28.1